LayoutTests/platform/ios/accessibility/list-with-dynamically-changing-content-expected.txt

This test ensures that lists that are initially empty but gain children later become part of the AX tree with the proper role.

#ul-list: Group
#ul-list: childrenCount 0
#aria-list: ApplicationGroup
#aria-list: childrenCount 0

Adding listitems to both lists.

#ul-list: List
#ul-list: childrenCount 1
#aria-list: List
#aria-list: childrenCount 1

Clearing listitems from both lists.

#ul-list: Group
#ul-list: childrenCount 0
#aria-list: ApplicationGroup
#aria-list: childrenCount 0

PASS successfullyParsed is true

TEST COMPLETE
